+++ src/lib/connection/RESTService.ts
@@ -6575,7 +6575,7 @@ Request Method: PUT
 
         return new Promise(function (resolve, reject) {
 
-            that.http.post(urlParams, that.getRequestHeader(""), data, 'text/csv; charset=utf-8').then(function (json) {
+            that.http.post(urlParams, that.getRequestHeader(""), data, undefined).then(function (json) {
                 that.logger.log("info", LOG_ID + "(updateCommandIdStatus) successfull");
                 that.logger.log("internal", LOG_ID + "(updateCommandIdStatus) REST result : ", json);
                 resolve(json);